+++ CHANGELOG.rst
@@ -1,6 +1,16 @@
 LATEST CHANGES
 ==============
 
+2025-02-13
+----------
+- Add image validity check support [SDK 5 required]
+  - Add new parameter 'general.enable_image_validity_check'
+  - Add new topic 'health_status/low_image_quality' to publish image quality status
+  - Add new topic 'health_status/low_lighting' to publish low light condition status
+  - Add new topic 'health_status/low_depth_reliability' to publish low depth quality status
+  - Add new topic 'health_status/low_motion_sensors_reliability' to publish low quality of inertial sensors status
+  - Set the Node Disgnostic to WARNING if any of the above conditions are detected
+
 2025-02-10
 ----------
 - Add new `OPTIMIZED` mode for `general.pub_resolution`